More Bot Changes
The bot isn't quite ready, but it's coming along nicely. This commit
mainly focuses on task execution. The gate file is almost complete, but
has no form of encrypted communication yet (same goes for the bot).

WARNING: This bot now contains features that would normally be found in
a typical bot, and may cause harm to your computer if used incorrectly!